+++ src/components/Lobby/index.module.scss
@@ -8,3 +8,51 @@
   
 }
 
+@import "#theme/theme";
+
+.lobbyRow {
+    background-color: $color-primary;
+    width: 600px;
+
+    border-left: 5px solid #9b2834;
+
+    details {
+        cursor: pointer;
+
+        span {
+            display: flex;
+            flex-direction: row;
+
+            
+        }
+
+        img {
+            width: 20px;
+            display: inline-flex;
+        }
+
+        summary {
+
+            display: flex;
+            flex-direction: row;
+            justify-content: space-between;
+
+        }
+
+        &[open] {
+            summary {
+                border-bottom: 1px solid darken($color-primary, 25%);
+        
+            }
+
+        }
+    }
+
+    .infoBox{
+        background-color: #7875b3;
+
+        border-bottom: 1px solid darken($color-primary, 25%);
+        }
+
+
+}
